src/requirements.txt
slackclient
flask
certifi
schedule